+++ app/routes/index.js
@@ -2,10 +2,11 @@
 
 import express from "express";
 import { router as mainRouter } from "./main.js";
+import { router as locationReverseRouter } from "./location/v1/reverse.js";
 
 const router = express.Router();
 
-// Main router.
 router.use(mainRouter);
+router.use(locationReverseRouter);
 
 export { router };