+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v1.0.0
 
 - [feat] Add core with multi language (spnish, english and german).
+- [feat] Add location/v1/reverse/longitude/latitude endpoint.